pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "fpdf>=1.7.2",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]
